api/requirements.txt
aiohttp
aioice
aiortc
av
cffi
cryptography
dlib
Flask
flask-cors
numpy
opencv-python
python-socketio
waitress
Werkzeug
psycopg[binary]
pyjwt
bcrypt
aiohttp_cors